[MISC] Renamed configuration to settings.py
